+++ src/main/java/net/olympiccode/vhackos/api/misc/Leaderboards.java
@@ -12,7 +12,7 @@ public interface Leaderboards {
     interface LeaderboardEntry {
        String getUsername();
        int getLevel();
-       long getExpPorcentage();
+       double getExpPorcentage();
     }
 
     interface TournamentEntry {
